requirements.txt
streamlit
pandas
plotly
google-cloud-bigquery
google-auth
db-dtypes
pyarrow
